Silence conversion warnings

+++ src/gui/Interface.cpp
@@ -2145,7 +2145,7 @@ void ArxGame::manageEditorControls() {
 					   && light->exist
 					   && !fartherThan(light->pos, player.pos, fMaxdist)
 					   && !(light->extras & EXTRAS_NO_IGNIT)
-					   && light->m_screenRect.toRect().contains(Vec2i(DANAEMouse))
+					   && light->m_screenRect.toRect().contains(Vec2f(DANAEMouse))
 					   && (COMBINE->ioflags & IO_ITEM)
 					) {
 						if((COMBINE == player.torch) || (COMBINE->_itemdata->LightValue == 1)) {
@@ -2191,7 +2191,7 @@ void ArxGame::manageEditorControls() {
 				   && light->exist
 				   && !fartherThan(light->pos, player.pos, fMaxdist)
 				   && !(light->extras & EXTRAS_NO_IGNIT)
-				   && light->m_screenRect.toRect().contains(Vec2i(DANAEMouse))
+				   && light->m_screenRect.toRect().contains(Vec2f(DANAEMouse))
 				) {
 					SpecialCursor = CURSOR_INTERACTION_ON;
 				}
